continue admin

+++ database/migrations/2022_10_25_191739_create_comments_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('post_id'); //When a post is deleted, also delete the comments in that post
-            $table->unsignedBigInteger('user_id');  //When a user is deleted, also delete the comments from that user
+            $table->unsignedInteger('post_id'); //When a post is deleted, also delete the comments in that post
+            $table->unsignedInteger('user_id');  //When a user is deleted, also delete the comments from that user
             $table->text('body');
             $table->timestamps();
         });
